+++ Pandora/Paradigm/Structure/Modification.hs
@@ -2,3 +2,4 @@ module Pandora.Paradigm.Structure.Modification (module Exports) where
 
 import Pandora.Paradigm.Structure.Modification.Prefixed as Exports
 import Pandora.Paradigm.Structure.Modification.Comprehension as Exports
+import Pandora.Paradigm.Structure.Modification.Combinative as Exports