+++ Pandora/Paradigm/Structure/Modification/Combinative.hs
@@ -1,0 +1,3 @@
+module Pandora.Paradigm.Structure.Modification.Combinative where
+
+type family Combinative (s :: * -> *) = (r :: * -> *) | r -> s