+++ CHANGELOG.md
@@ -432,3 +432,4 @@
 * Define `lookup` method in `Morphable` ability module
 * Remove `binary` expression from `Binary` module
 * Define `Morphed` type synonymous as constraint kind
+* Define `Combinative` type family